case size_t to double

+++ external/nsfd/src/nsfd/grid/axis.hpp
@@ -21,9 +21,9 @@ class Axis {
 
   Axis() : length_{0}, frac_{0}, del_{0}, max{0} {};
   Axis(double length, size_t max, double frac)
-      : length_{length}, frac_{frac}, del_{length / max}, max{max} {};
+      : length_{length}, frac_{frac}, del_{length / static_cast<double>(max)}, max{max} {};
   Axis(double length, size_t max)
-      : length_{length}, frac_{0.5}, del_{length / max}, max{max} {};
+      : length_{length}, frac_{0.5}, del_{length / static_cast<double>(max)}, max{max} {};
 
   double operator[](size_t i) const {
     return del_ * (static_cast<double>(i) - frac_);
